learning

+++ 0x00-hello_world/6-size.c
@@ -11,7 +11,7 @@ char c;
 long int d;
 
 float f;
-printf("Size of a char: %zu bytes(s)\n", (unsigned long)sizeof(c));
+printf("Size of a char: %zu bytes(s)\n", sizeof(c));
 printf("Size of an int: %zu bytes(s)\n", (unsigned long)sizeof(i));
 printf("Size of a long int: %zu bytes(s)\n", sizeof(d));
 printf("Size of a long long int: %lu bytes(s)\n", sizeof(long long int));
